+++ WorldEditCommands/Object/DataAutoComplete.cs
@@ -2,19 +2,25 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using System.Reflection.Emit;
 using BepInEx.Bootstrap;
+using HarmonyLib;
 using ServerDevcommands;
-using Service;
 using UnityEngine;
 
 namespace WorldEditCommands;
 
 public class DataAutoComplete
 {
-  private static Dictionary<string, Dictionary<string, Type>>? fields;
-  public static Dictionary<string, Dictionary<string, Type>> Fields => fields ??= LoadFields();
-  private static List<string>? components;
-  private static List<string> Components => components ??= [.. Fields.Keys];
+  public static Dictionary<string, Dictionary<string, Type>> Fields = [];
+  private static List<string> Components = [];
+
+  public static void Init()
+  {
+    // Easier to always load everything at start up, doesn't take long anyways.
+    Fields = LoadFields();
+    Components = [.. Fields.Keys];
+  }
   private static readonly HashSet<Type> ValidTypes = [
     typeof(string),
     typeof(int),
@@ -23,22 +29,29 @@ public class DataAutoComplete
     typeof(Vector3),
     typeof(GameObject)
   ];
-  public static Dictionary<string, Dictionary<string, Type>> LoadFields()
+  private static readonly HashSet<Type> ValidTweakTypes = [
+    ..ValidTypes,
+    typeof(ItemDrop),
+    typeof(Character.Faction),
+    typeof(EffectList)
+  ];
+  private static Dictionary<string, Dictionary<string, Type>> LoadFields()
   {
     Dictionary<string, Dictionary<string, Type>> fields = [];
     List<Assembly> assemblies = [Assembly.GetAssembly(typeof(ZNetView)), .. Chainloader.PluginInfos.Values.Where(p => p.Instance != null).Select(p => p.Instance.GetType().Assembly)];
     var assembly = Assembly.GetAssembly(typeof(ZNetView));
     var baseType = typeof(MonoBehaviour);
     var types = assemblies.SelectMany(s => s.GetTypes()).Where(baseType.IsAssignableFrom);
-
+    var valid = WorldEditCommands.IsTweaks ? ValidTweakTypes : ValidTypes;
     foreach (var type in types)
     {
       if (!fields.ContainsKey(type.Name))
         fields[type.Name] = [];
       foreach (var fieldInfo in type.GetFields(BindingFlags.Instance | BindingFlags.Public))
       {
-        if (!ValidTypes.Contains(fieldInfo.FieldType)) continue;
+        if (!valid.Contains(fieldInfo.FieldType)) continue;
         fields[type.Name][fieldInfo.Name] = fieldInfo.FieldType;
+        PreventStripping.AddKey(type, fieldInfo);
       }
       if (fields[type.Name].Count == 0)
         fields.Remove(type.Name);
@@ -101,6 +114,7 @@ public class DataAutoComplete
   }
   public static List<string> GetTypes(int index)
   {
+    var tweaks = WorldEditCommands.IsTweaks;
     var command = GetInput();
     var prefab = PrefabFromCommand(command);
     var component = ComponentFromCommand(prefab, command);
@@ -112,8 +126,33 @@ public class DataAutoComplete
     if (type == typeof(float)) return ["Decimal"];
     if (type == typeof(bool)) return ["true", "false"];
     if (type == typeof(Vector3)) return ServerDevcommands.ParameterInfo.XZY("Field", index);
-    if (type == typeof(GameObject)) return ServerDevcommands.ParameterInfo.Ids;
+    if (type == typeof(GameObject)) return tweaks ? GetIdsOrTransforms(prefab, component, field) : ServerDevcommands.ParameterInfo.Ids;
+    if (tweaks && type == typeof(ItemDrop)) return ServerDevcommands.ParameterInfo.ItemIds;
+    if (tweaks && type == typeof(Character.Faction)) return [.. Enum.GetNames(typeof(Character.Faction))];
+    if (tweaks && type == typeof(EffectList)) return ServerDevcommands.ParameterInfo.Ids;
     return [];
+  }
+  private static List<string> GetIdsOrTransforms(string prefab, string component, string field)
+  {
+    var obj = ZNetScene.instance.GetPrefab(prefab);
+    if (!obj) return ServerDevcommands.ParameterInfo.Ids;
+    var c = obj.GetComponents<MonoBehaviour>().FirstOrDefault(c => c.GetType().Name == component);
+    if (c == null) return ServerDevcommands.ParameterInfo.Ids;
+    var f = c.GetType().GetField(RealField(component, field));
+    if (f == null || f.FieldType != typeof(GameObject)) return ServerDevcommands.ParameterInfo.Ids;
+    var go = (GameObject)f.GetValue(c);
+    if (go && ZNetScene.instance.GetPrefab(go.name) == go) return ServerDevcommands.ParameterInfo.Ids;
+    List<string> trs = [];
+    GetChild("/", obj.transform, trs);
+    return trs;
+  }
+  private static void GetChild(string prefix, Transform tr, List<string> trs)
+  {
+    foreach (Transform child in tr)
+    {
+      trs.Add($"{prefix}{child.name}");
+      GetChild($"{prefix}{child.name}/", child, trs);
+    }
   }
   private static string FieldFromCommand(string component, string command)
   {
@@ -141,4 +180,40 @@ public class DataAutoComplete
     if (secondaryField != null) return secondaryField;
     return field;
   }
+}
+
+[HarmonyPatch(typeof(ZDO))]
+public class PreventStripping
+{
+  private static readonly HashSet<int> FieldZdoKeys = [];
+
+  public static void AddKey(Type type, FieldInfo fieldInfo)
+  {
+    if (fieldInfo.FieldType == typeof(string) || fieldInfo.FieldType == typeof(int) || fieldInfo.FieldType == typeof(bool))
+    {
+      var key = $"{type.Name}.{fieldInfo.Name}";
+      FieldZdoKeys.Add(key.GetStableHashCode());
+    }
+  }
+
+  // Transpiling is more tricky but should be more performant.
+  [HarmonyPatch(nameof(ZDO.Strip), typeof(int), typeof(string)), HarmonyTranspiler]
+  static IEnumerable<CodeInstruction> String(IEnumerable<CodeInstruction> instructions, ILGenerator generator) => Transpile(instructions, generator);
+  [HarmonyPatch(nameof(ZDO.Strip), typeof(int), typeof(int)), HarmonyTranspiler]
+  static IEnumerable<CodeInstruction> Int(IEnumerable<CodeInstruction> instructions, ILGenerator generator) => Transpile(instructions, generator);
+
+  private static IEnumerable<CodeInstruction> Transpile(IEnumerable<CodeInstruction> instructions, ILGenerator generator)
+  {
+    return new CodeMatcher(instructions, generator)
+      .CreateLabelAt(0, out var label)
+      .Start()
+      .InsertAndAdvance(new CodeInstruction(OpCodes.Ldsfld, AccessTools.Field(typeof(PreventStripping), nameof(FieldZdoKeys))))
+      .InsertAndAdvance(new CodeInstruction(OpCodes.Ldarg_0))
+      .InsertAndAdvance(new CodeInstruction(OpCodes.Call, AccessTools.Method(typeof(HashSet<int>), nameof(HashSet<int>.Contains))))
+      .InsertAndAdvance(new CodeInstruction(OpCodes.Brfalse, label)) // Jump to the original code, otherwise return false.
+      .InsertAndAdvance(new CodeInstruction(OpCodes.Ldc_I4_0))
+      .InsertAndAdvance(new CodeInstruction(OpCodes.Ret))
+    .InstructionEnumeration();
+  }
+
 }